% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_identification.R
\name{calc_alk_efficiency}
\alias{calc_alk_efficiency}
\title{计算烷基化效率}
\usage{
calc_alk_efficiency(psm_df)
}
\arguments{
\item{psm_df}{PSM 水平 data.frame (需含 Peptide 和 Assigned Modifications 列)}
}
\value{
百分比数值
}
\description{
计算烷基化效率
}
