% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_identification.R
\name{calc_cys_percent}
\alias{calc_cys_percent}
\title{计算含 Cys 的 PSM 百分比}
\usage{
calc_cys_percent(psm_df)
}
\arguments{
\item{psm_df}{PSM 水平 data.frame (需含 Peptide 列)}
}
\value{
百分比数值
}
\description{
计算含 Cys 的 PSM 百分比
}
