% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_evaluation.R, R/qc_identification.R
\name{calc_gravy}
\alias{calc_gravy}
\title{计算 GRAVY 疏水性指数}
\usage{
calc_gravy(seqs)

calc_gravy(seqs)
}
\arguments{
\item{seqs}{Character vector of peptide sequences}
}
\value{
GRAVY 数值向量

Numeric vector of GRAVY scores
}
\description{
Kyte-Doolittle hydropathy index average.
}
