% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{calc_lod}
\alias{calc_lod}
\title{\u8ba1\u7b97 LOD\uff08\u6700\u4f4e\u53ef\u68c0\u51fa\u6d53\u5ea6\uff09}
\usage{
calc_lod(bsa_df, min_peptides = 1)
}
\arguments{
\item{bsa_df}{AlphaTims BSA metrics data.frame}

\item{min_peptides}{\u5b9a\u4e49\u68c0\u51fa\u6240\u9700\u6700\u5c11\u80bd\u6570 (default: 1)}
}
\value{
data.frame\uff0c\u542b\u5206\u7ec4\u548c LOD \u5c45\u5ea6
}
\description{
\u57fa\u4e8e AlphaTims \u7684 BSA \u7279\u5f81\u80bd\u68c0\u51fa\u6570\u636e\uff0c\u8ba1\u7b97\u6bcf\u6839\u8272\u8c31\u67f1
\u7684 LOD\uff08\u5b9a\u4e49\u4e3a\uff1a\u5728 2 \u4e2a\u91cd\u590d\u4e2d\u5747\u68c0\u51fa\u81f3\u5c11\u4e00\u4e2a\u7279\u5f81\u80bd\u7684\u6700\u4f4e\u6d53\u5ea6\uff09\u3002
}
