% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_evaluation.R
\name{calc_missed_cleavage}
\alias{calc_missed_cleavage}
\title{计算 Missed Cleavage 数量}
\usage{
calc_missed_cleavage(seqs)
}
\arguments{
\item{seqs}{肽段序列字符串向量}
}
\value{
整数向量
}
\description{
计算 Missed Cleavage 数量
}
