% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_evaluation.R
\name{calc_pI}
\alias{calc_pI}
\title{计算等电点 pI (Bjellqvist scale)}
\usage{
calc_pI(seq)
}
\arguments{
\item{seq}{氨基酸序列字符串}
}
\value{
pI 数值
}
\description{
计算等电点 pI (Bjellqvist scale)
}
\examples{
calc_pI("ACDEFGHIKLMNPQRSTVWY")
}
