% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_identification.R
\name{calc_pI_seq}
\alias{calc_pI_seq}
\title{Calculate peptide isoelectric point (pI)}
\usage{
calc_pI_seq(seqs)
}
\arguments{
\item{seqs}{Character vector of peptide sequences}
}
\value{
Numeric vector of pI values
}
\description{
Bjellqvist scale bisection algorithm.
}
