% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_identification.R
\name{calc_zero_miss}
\alias{calc_zero_miss}
\title{计算 0-Miss Cleavage 比例}
\usage{
calc_zero_miss(psm_df)
}
\arguments{
\item{psm_df}{PSM 水平 data.frame (需含 Peptide 和 Number of Missed Cleavages 列)}
}
\value{
百分比数值
}
\description{
从 PSM 数据中计算去冗余后 0 漏切比例。
}
