% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_extraction.R
\name{.check_python_alphatims}
\alias{.check_python_alphatims}
\title{Check Python and alphatims availability}
\usage{
.check_python_alphatims()
}
\description{
Check Python and alphatims availability
}
\keyword{internal}
