% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_spectronaut.R
\name{.detect_spectronaut_header}
\alias{.detect_spectronaut_header}
\title{通过读取表头判断是否为 Spectronaut 文件}
\usage{
.detect_spectronaut_header(fpath)
}
\description{
通过读取表头判断是否为 Spectronaut 文件
}
\keyword{internal}
