% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.extract_diff_genes}
\alias{.extract_diff_genes}
\title{Extract gene symbols from differential analysis result}
\usage{
.extract_diff_genes(diff_result)
}
\description{
Extract gene symbols from differential analysis result
}
\keyword{internal}
