% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.extract_group_genes}
\alias{.extract_group_genes}
\title{Extract gene symbols from MsDataSet by group}
\usage{
.extract_group_genes(ms_data, group_info)
}
\description{
Extract gene symbols from MsDataSet by group
}
\keyword{internal}
