% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plasma.R
\name{.img_to_b64}
\alias{.img_to_b64}
\title{Encode PNG file as base64 img tag}
\usage{
.img_to_b64(path, width = "85\%", caption = "")
}
\arguments{
\item{path}{Path to PNG file}

\item{width}{CSS width (default "85\%")}

\item{caption}{Figure caption}
}
\value{
HTML string
}
\description{
Encode PNG file as base64 img tag
}
\keyword{internal}
