% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contamination.R
\name{.match_markers_gene}
\alias{.match_markers_gene}
\title{Match marker proteins by gene name}
\usage{
.match_markers_gene(genes, pinfo)
}
\arguments{
\item{genes}{Character vector of gene names to match}

\item{pinfo}{data.frame with "Gene" column}
}
\value{
Integer vector of row indices
}
\description{
Match marker proteins by gene name
}
\keyword{internal}
