% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.msProteomiX_env}
\alias{.msProteomiX_env}
\title{msProteomiX 包内部环境}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.msProteomiX_env
}
\description{
用于在 RStudio Source 模式下跨脚本传递数据
}
\keyword{internal}
