% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.run_correlation_network}
\alias{.run_correlation_network}
\title{Build co-expression network from protein abundance correlation (internal)}
\usage{
.run_correlation_network(diff_result, ms_data, cor_threshold = 0.8)
}
\description{
Build co-expression network from protein abundance correlation (internal)
}
\keyword{internal}
