% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.symbol_to_entrez}
\alias{.symbol_to_entrez}
\title{Convert gene symbols to Entrez IDs}
\usage{
.symbol_to_entrez(gene_symbols, org_db)
}
\description{
Convert gene symbols to Entrez IDs
}
\keyword{internal}
