% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{正则转义辅助函数}
\usage{
escape_regex(s)
}
\arguments{
\item{s}{要转义的字符串}
}
\value{
转义后的字符串
}
\description{
转义字符串中的正则特殊字符，保护样本名中的 +, ., (, ) 等
}
\examples{
escape_regex("Sample+1 (Rep.2)")
}
