% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{import_alphatims_results}
\alias{import_alphatims_results}
\title{\u5bfc\u5165 AlphaTims \u5206\u6790\u7ed3\u679c}
\usage{
import_alphatims_results(
  bsa_csv = "bsa_metrics.csv",
  irt_csv = "irt_metrics.csv"
)
}
\arguments{
\item{bsa_csv}{BSA \u7279\u5f81\u80bd CSV \u8def\u5f84 (bsa_metrics.csv)}

\item{irt_csv}{iRT \u7387\u5b9a CSV \u8def\u5f84 (irt_metrics.csv)}
}
\value{
\u5217\u8868\uff0c\u542b\u4e24\u4e2a data.frame\uff1a$bsa \u548c $irt
}
\description{
\u8bfb\u53d6 AlphaTims Python \u811a\u672c\u8f93\u51fa\u7684 BSA \u7279\u5f81\u80bd\u548c iRT \u6307\u6807 CSV\uff0c
\u8fd4\u56de\u5e26\u5e26\u5206\u7ec4\u5143\u4fe1\u606f\u7684\u6807\u51c6\u5316 data.frame\u3002
}
