% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{merge_fragpipe_runs}
\alias{merge_fragpipe_runs}
\title{\u5408\u5e76\u591a\u7ec4 FragPipe \u641c\u5e93\u7ed3\u679c}
\usage{
merge_fragpipe_runs(
  bsa_dir = "fragpipe_out/BSA_all",
  hela_dir = "fragpipe_out/HeLa_all"
)
}
\arguments{
\item{bsa_dir}{BSA FragPipe \u8f93\u51fa\u76ee\u5f55 (fragpipe_out/BSA_all)}

\item{hela_dir}{HeLa FragPipe \u8f93\u51fa\u76ee\u5f55 (fragpipe_out/HeLa_all)}
}
\value{
\u5217\u8868\uff0c\u542b $bsa (MsDataSet) \u548c $hela (MsDataSet)
}
\description{
\u5c06 BSA \u548c HeLa \u4e24\u4e2a\u72ec\u7acb\u7684 FragPipe \u641c\u5e93\u76ee\u5f55\u5206\u522b\u89e3\u6790\u5e76\u5b58\u5165\u5217\u8868\u3002
}
