% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_MsDataSet}
\alias{new_MsDataSet}
\title{创建 MsDataSet 对象}
\usage{
new_MsDataSet(
  proteins = data.frame(),
  peptides = data.frame(),
  psms = data.frame(),
  ions = data.frame(),
  sample_names = character(),
  protein_info = data.frame(),
  engine = "unknown",
  quant_type = "LFQ"
)
}
\arguments{
\item{proteins}{蛋白水平 data.frame}

\item{peptides}{肽段水平 data.frame (可选)}

\item{psms}{PSM 水平 data.frame (可选)}

\item{ions}{离子水平 data.frame (可选)}

\item{sample_names}{样本名向量}

\item{protein_info}{蛋白注释信息 data.frame}

\item{engine}{搜库引擎名称}

\item{quant_type}{定量方式}
}
\value{
MsDataSet 对象
}
\description{
创建 MsDataSet 对象
}
\keyword{internal}
