% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_fragpipe.R
\name{parse_fragpipe}
\alias{parse_fragpipe}
\title{解析 FragPipe 搜库结果}
\usage{
parse_fragpipe(path)
}
\arguments{
\item{path}{文件路径或包含 FragPipe 结果的目录}
}
\value{
MsDataSet 对象
}
\description{
读取 FragPipe 输出的 combined_protein.tsv (及可选的 combined_peptide.tsv,
combined_ion.tsv, psm.tsv) 文件，返回 MsDataSet 对象。
}
