% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_apms.R
\name{plot_apms_dotplot}
\alias{plot_apms_dotplot}
\title{AP-MS Dot Plot (核心可视化)}
\usage{
plot_apms_dotplot(rel_abundance, sig_proteins = NULL, project_name = "APMS")
}
\arguments{
\item{rel_abundance}{calc_relative_abundance() 返回的 data.frame}

\item{sig_proteins}{显著蛋白 ID 向量 (NULL = 使用全部)}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
经典 AP-MS 气泡图：X 轴为时间点/条件，Y 轴为基因名，
气泡大小表示相对丰度，颜色表示最大丰度所在的组。
}
