% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_apms.R
\name{plot_apms_heatmap}
\alias{plot_apms_heatmap}
\title{AP-MS 热图}
\usage{
plot_apms_heatmap(rel_abundance, sig_proteins = NULL, project_name = "APMS")
}
\arguments{
\item{rel_abundance}{calc_relative_abundance() 返回的 data.frame}

\item{sig_proteins}{显著蛋白 ID 向量 (NULL = 使用全部)}

\item{project_name}{项目名}
}
\value{
pheatmap 对象
}
\description{
显著蛋白的相对丰度热图 (用 pheatmap)。
行按峰值时间点分组排序。
}
