% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{绘制序列覆盖度箱线图}
\usage{
plot_coverage(coverage_df, output_dir = "output", project_name = "Project")
}
\arguments{
\item{coverage_df}{calc_coverage() 返回的 data.frame}

\item{output_dir}{输出目录}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
绘制序列覆盖度箱线图
}
