% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_cv.R
\name{plot_cv_boxplot}
\alias{plot_cv_boxplot}
\title{绘制 CV 变异系数箱线图}
\usage{
plot_cv_boxplot(
  ms_data,
  group_info,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet 对象}

\item{group_info}{分组信息 data.frame}

\item{output_dir}{输出目录}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
绘制 CV 变异系数箱线图
}
