% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_go_bubble}
\alias{plot_go_bubble}
\title{绘制 GO 富集气泡图}
\usage{
plot_go_bubble(go_df, output_dir = "output", project_name = "Project")
}
\arguments{
\item{go_df}{run_go_enrichment() 返回的 data.frame}

\item{output_dir}{输出目录}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
绘制 GO 富集气泡图
}
