% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{plot_hela_loading_curve}
\alias{plot_hela_loading_curve}
\title{\u7ed8\u5236 HeLa \u4e0a\u6837\u91cf vs \u86cb\u767d\u8d28\u6570\u66f2\u7ebf}
\usage{
plot_hela_loading_curve(
  hela_ms,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{hela_ms}{MsDataSet \u5bf9\u8c61 (FragPipe HeLa \u641c\u5e93\u7ed3\u679c)}

\item{output_dir}{\u8f93\u51fa\u76ee\u5f55}

\item{project_name}{\u9879\u76ee\u540d}
}
\value{
ggplot \u5bf9\u8c61
}
\description{
\u4ee5\u4e0a\u6837\u91cf\u4e3a X \u8f74\uff0c\u68c0\u5230\u7684\u86cb\u767d\u8d28\u6570/\u80bd\u6bb5\u6570\u4e3a Y \u8f74\uff0c\u5bf9\u6bd4 ColA vs ColB\u3002
}
