% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{plot_irt_column_compare}
\alias{plot_irt_column_compare}
\title{\u7ed8\u5236 iRT \u8272\u8c71\u67f1\u5bf9\u6bd4\u56fe\uff083-panel\uff09}
\usage{
plot_irt_column_compare(
  irt_df,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{irt_df}{AlphaTims iRT metrics data.frame}

\item{output_dir}{\u8f93\u51fa\u76ee\u5f55}

\item{project_name}{\u9879\u76ee\u540d}
}
\value{
ggplot \u5bf9\u8c61 (\u7ec4\u5408\u56fe)
}
\description{
\u9762\u677f 1\uff1a\u5404\u6d53\u5ea6\u4e0b iRT \u5e73\u5747\u5cf0\u9762\u79ef\uff08ColA vs ColB\uff09
\u9762\u677f 2\uff1a\u5404 iRT \u80bd\u6bb5\u7684\u5e73\u5747 FWHM\uff08\u5cf0\u5bbd\uff09
\u9762\u677f 3\uff1a\u91cd\u590d\u95f4 CV\%\uff08\u5cf0\u9762\u79ef\uff09
}
