% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_kegg_bar}
\alias{plot_kegg_bar}
\title{Plot KEGG enrichment bar chart}
\usage{
plot_kegg_bar(kegg_df, output_dir = "output", project_name = "Project")
}
\arguments{
\item{kegg_df}{run_kegg_enrichment() result data.frame}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
ggplot object
}
\description{
Plot KEGG enrichment bar chart
}
