% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{plot_linearity}
\alias{plot_linearity}
\title{\u7ed8\u5236 BSA \u5f3a\u5ea6\u7ebf\u6027\u56fe\uff08log-log\uff09}
\usage{
plot_linearity(bsa_df, output_dir = "output", project_name = "Project")
}
\arguments{
\item{bsa_df}{AlphaTims BSA metrics data.frame}

\item{output_dir}{\u8f93\u51fa\u76ee\u5f55}

\item{project_name}{\u9879\u76ee\u540d}
}
\value{
ggplot \u5bf9\u8c61
}
\description{
\u5bf9\u6bcf\u6839\u8272\u8c71\u67f1\u548c\u6bcf\u6761 BSA \u80bd\u6bb5\uff0c\u7ed8\u5236 log2(\u5f3a\u5ea6) ~ log2(\u6d53\u5ea6)
\u7684\u56de\u5f52\u76f4\u7ebf\u548c R\u00b2\u3002
}
