% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{plot_lod_curve}
\alias{plot_lod_curve}
\title{\u7ed8\u5236 BSA LOD \u66f2\u7ebf}
\usage{
plot_lod_curve(
  bsa_df,
  output_dir = "output",
  project_name = "Project",
  peptide_type = "bsa"
)
}
\arguments{
\item{bsa_df}{AlphaTims BSA metrics data.frame}

\item{output_dir}{\u8f93\u51fa\u76ee\u5f55}

\item{project_name}{\u9879\u76ee\u540d}

\item{peptide_type}{\u5206\u6790\u5bf9\u8c61 (\u53ef\u9009 "bsa" \u6216 "irt")}
}
\value{
ggplot \u5bf9\u8c61
}
\description{
\u4ee5\u6d53\u5ea6\u4e3a X \u8f74\uff0c\u68c0\u51fa\u80bd\u6bb5\u6570\u4e3a Y \u8f74\uff0c\u5bf9\u6bd4 ColA vs ColB\u3002
}
