% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_apms.R
\name{plot_mfuzz}
\alias{plot_mfuzz}
\title{Mfuzz 聚类结果可视化}
\usage{
plot_mfuzz(mfuzz_result, project_name = "APMS")
}
\arguments{
\item{mfuzz_result}{run_mfuzz_cluster() 返回的 list}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
绘制 Mfuzz 软聚类的中心折线图和成员热图。
}
