% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_qc.R
\name{plot_qc_cys_alkylation}
\alias{plot_qc_cys_alkylation}
\title{Plot Cys-containing peptide ratio and alkylation efficiency}
\usage{
plot_qc_cys_alkylation(
  ms_data,
  group_info = NULL,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet object}

\item{group_info}{Group info data.frame}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
ggplot object or NULL
}
\description{
Plot Cys-containing peptide ratio and alkylation efficiency
}
