% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_qc.R
\name{plot_qc_mass_error}
\alias{plot_qc_mass_error}
\title{Plot mass error (ppm) distribution}
\usage{
plot_qc_mass_error(
  ms_data,
  group_info = NULL,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet object}

\item{group_info}{Group info data.frame}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
ggplot object or NULL
}
\description{
Density plot of mass accuracy in ppm.
Requires "Calibrated Observed Mass" and "Calculated Peptide Mass" columns.
Currently available for FragPipe only.
}
