% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_qc.R
\name{plot_qc_modification}
\alias{plot_qc_modification}
\title{Plot modification type distribution}
\usage{
plot_qc_modification(
  ms_data,
  top_n = 10,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet object}

\item{top_n}{Number of top modifications to show}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
ggplot object or NULL
}
\description{
Plot modification type distribution
}
