% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_qc.R
\name{plot_qc_panel}
\alias{plot_qc_panel}
\title{Generate QC evaluation panel (all QC plots)}
\usage{
plot_qc_panel(
  ms_data,
  group_info = NULL,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet object}

\item{group_info}{Group info data.frame (NULL = overall)}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
List of generated ggplot objects (invisible)
}
\description{
Generate QC evaluation panel (all QC plots)
}
