% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_reactome_bar}
\alias{plot_reactome_bar}
\title{Plot Reactome enrichment bar chart}
\usage{
plot_reactome_bar(reactome_df, output_dir = "output", project_name = "Project")
}
\arguments{
\item{reactome_df}{run_reactome_enrichment() result data.frame}

\item{output_dir}{Output directory}

\item{project_name}{Project name}
}
\value{
ggplot object
}
\description{
Plot Reactome enrichment bar chart
}
