% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_venn.R
\name{plot_upset}
\alias{plot_upset}
\title{绘制 UpSet 图}
\usage{
plot_upset(
  ms_data,
  group_info,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet 对象}

\item{group_info}{分组信息 data.frame}

\item{output_dir}{输出目录}

\item{project_name}{项目名}
}
\description{
绘制 UpSet 图
}
