% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{绘制 Venn 维恩图}
\usage{
plot_venn(
  ms_data,
  group_info,
  selected_groups = NULL,
  output_dir = "output",
  project_name = "Project"
)
}
\arguments{
\item{ms_data}{MsDataSet 对象}

\item{group_info}{分组信息 data.frame}

\item{selected_groups}{选择绘制的组名 (NULL 表示全部，最多 4 组)}

\item{output_dir}{输出目录}

\item{project_name}{项目名}
}
\value{
ggplot 对象
}
\description{
绘制 Venn 维恩图
}
