% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apms.R
\name{run_mfuzz_cluster}
\alias{run_mfuzz_cluster}
\title{Mfuzz 软聚类}
\usage{
run_mfuzz_cluster(
  ms_data,
  group_info,
  sig_proteins,
  n_clusters = 4,
  min_valid = 2
)
}
\arguments{
\item{ms_data}{MsDataSet 对象}

\item{group_info}{分组信息}

\item{sig_proteins}{显著蛋白 ID 向量 (来自 ANOVA 结果)}

\item{n_clusters}{聚类数目 (默认 4)}

\item{min_valid}{每组最少有效值数 (默认 2)}
}
\value{
list(eset=ExpressionSet, cl=Mfuzz聚类结果, membership=成员矩阵)
}
\description{
对显著蛋白做 Mfuzz 软聚类，按时序模式分组。
需要安装 Bioconductor 包 Mfuzz。
}
