% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_reactome_enrichment}
\alias{run_reactome_enrichment}
\title{Run Reactome pathway enrichment analysis}
\usage{
run_reactome_enrichment(
  ms_data = NULL,
  group_info = NULL,
  diff_result = NULL,
  org_db = "org.Hs.eg.db",
  organism = "human",
  top_n = 20
)
}
\arguments{
\item{ms_data}{MsDataSet object (used when diff_result is NULL)}

\item{group_info}{Group info data.frame (used when diff_result is NULL)}

\item{diff_result}{Optional. Output from run_diff_analysis().}

\item{org_db}{Annotation package name (default "org.Hs.eg.db")}

\item{organism}{Reactome organism (default "human")}

\item{top_n}{Number of top results to keep}
}
\value{
data.frame with enrichment results
}
\description{
Run Reactome pathway enrichment analysis
}
