% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.classify_output_files}
\alias{.classify_output_files}
\title{Classify output files by analysis type}
\usage{
.classify_output_files(pdf_files)
}
\description{
Classify output files by analysis type
}
\keyword{internal}
