% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_fragpipe.R
\name{.extract_fp_quant_cols}
\alias{.extract_fp_quant_cols}
\title{提取 FragPipe 定量列名和样本名}
\usage{
.extract_fp_quant_cols(raw_cols)
}
\description{
FragPipe combined_protein.tsv 的定量列格式为:
"SampleName MaxLFQ Intensity" (样本名在前，后缀固定)
}
\keyword{internal}
