% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.generate_rmd_template}
\alias{.generate_rmd_template}
\title{Generate RMarkdown template dynamically}
\usage{
.generate_rmd_template(output_dir, project_name, pdf_files, csv_files)
}
\description{
Generate RMarkdown template dynamically
}
\keyword{internal}
