% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.msProteomiX_env}
\alias{.msProteomiX_env}
\title{msProteomiX shared environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.msProteomiX_env
}
\description{
Used to pass data between scripts in RStudio Source mode
}
\keyword{datasets}
