% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_go_enrichment}
\alias{run_go_enrichment}
\title{运行 GO 富集分析 (Cellular Component)}
\usage{
run_go_enrichment(
  ms_data,
  group_info,
  org_db = "org.Hs.eg.db",
  ont = "CC",
  top_n = 10
)
}
\arguments{
\item{ms_data}{MsDataSet 对象 (需含 peptides 数据)}

\item{group_info}{分组信息 data.frame}

\item{org_db}{物种注释包名 (默认 "org.Hs.eg.db" 人类)}

\item{ont}{GO Ontology: "CC" (Cellular Component), "BP", "MF"}

\item{top_n}{每组保留前 N 个结果}
}
\value{
data.frame
}
\description{
运行 GO 富集分析 (Cellular Component)
}
